+++ app/routes/components/IbsBase/changelog.codegen.ts
@@ -3,4 +3,6 @@
 
 export const changelog = ''
   + '- 2025-12-07: Initial version.\n'
+  + '- 2025-12-16: Use `min-height` instead of `height` to let components built on\n'
+  + '  top of it to have a flexible height.\n'
   ;