+++ src/components/IbsBase/ibs-base-left-section.tsx
@@ -1,6 +1,8 @@
 import classNames from "classnames";
 import type { ComponentProps, FC } from "react";
 
+import { useJss } from "@/components/CopyUiProvider";
+
 import styles from "./ibs-base-left-section.module.scss";
 
 type IbsBaseLeftSectionProps = ComponentProps<"div">;
@@ -8,8 +10,17 @@ type IbsBaseLeftSectionProps = ComponentProps<"div">;
 const IbsBaseLeftSection: FC<IbsBaseLeftSectionProps> = (props) => {
   const { className, children, ...rest } = props;
 
+  const jss = useJss();
+  const stx = jss.hash({
+    "--ibsBaseLeftSection-w": "calc(var(--ibsBase-minH) - 2px)",
+    "--ibsBaseLeftSection-h": "calc(var(--ibsBase-minH) - 2px)",
+  });
+
   return (
-    <div className={classNames(styles.ibsBaseLeftSection, className)} {...rest}>
+    <div
+      className={classNames(styles.ibsBaseLeftSection, stx, className)}
+      {...rest}
+    >
       {children}
     </div>
   );