+++ src/components/IbsBase/ibs-base-right-section.tsx
@@ -1,6 +1,8 @@
 import classNames from "classnames";
 import type { ComponentProps, FC } from "react";
 
+import { useJss } from "@/components/CopyUiProvider";
+
 import styles from "./ibs-base-right-section.module.scss";
 
 type IbsBaseRightSectionProps = ComponentProps<"div">;
@@ -8,9 +10,15 @@ type IbsBaseRightSectionProps = ComponentProps<"div">;
 const IbsBaseRightSection: FC<IbsBaseRightSectionProps> = (props) => {
   const { className, children, ...rest } = props;
 
+  const jss = useJss();
+  const stx = jss.hash({
+    "--ibsBaseRightSection-w": "calc(var(--ibsBase-minH) - 2px)",
+    "--ibsBaseRightSection-h": "calc(var(--ibsBase-minH) - 2px)",
+  });
+
   return (
     <div
-      className={classNames(styles.ibsBaseRightSection, className)}
+      className={classNames(styles.ibsBaseRightSection, stx, className)}
       {...rest}
     >
       {children}