+++ src/components/IbsBase/index.ts
@@ -1,4 +1,4 @@
-// IbsBase from copy-ui @ 2025-11-28.
+// IbsBase from copy-ui @ 2025-12-16.
 
 export { IbsBase } from "./ibs-base";
 export { extractStylesProps } from "./style";