+++ src/components/IbsBase/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-12-07: Initial version.
+- 2025-12-16: Use `min-height` instead of `height` to let components built on
+  top of it to have a flexible height.